requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.15.2
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
pwdlib[argon2]==0.2.1
python-multipart==0.0.20
email-validator==2.2.0
psycopg[binary]==3.2.13
pytest==8.3.5
httpx==0.28.1
ruff==0.11.2

